model/baselinewithfacenet/requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
astunparse==1.6.3
attrs==24.3.0
av==14.0.1
backcall==0.2.0
blinker==1.6.2
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.5.0
certifi @ file:///croot/certifi_1734473278428/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.3
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.6.6
decorator==5.1.1
dill==0.3.9
dlib==19.24.6
executing==1.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
facenet-pytorch==2.6.0
filelock @ file:///croot/filelock_1700591183607/work
Flask==2.2.3
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.12.0
gast==0.4.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
google-auth==2.37.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
idna @ file:///croot/idna_1714398848350/work
importlib-metadata==6.0.0
importlib_resources==6.4.5
imutils==0.5.4
ipykernel==6.21.3
ipython==8.11.0
itsdangerous==2.1.2
jax==0.4.30
jaxlib==0.4.30
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.3
jupyter_core==5.2.0
keras==3.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.9.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.20
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
ml-dtypes==0.4.1
mpmath @ file:///croot/mpmath_1690848262763/work
namex==0.0.8
nest-asyncio==1.5.6
networkx @ file:///croot/networkx_1717597493534/work
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==23.0
pandas==2.2.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.3.6
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.14.0
pyparsing==3.2.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
pytz==2024.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==25.0.0
requests @ file:///croot/requests_1730999120400/work
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.19.2
six==1.16.0
sounddevice==0.5.1
stack-data==0.6.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.18.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
thop==0.1.1.post2209072238
torch==2.2.2
torchaudio==2.5.1
torchvision==0.17.2
tornado==6.2
tqdm==4.67.1
traitlets==5.14.3
triton==2.2.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2024.2
ultralytics==8.3.52
ultralytics-thop==2.0.13
urllib3 @ file:///croot/urllib3_1727769808118/work
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.17.0
zipp==3.15.0
